Edit ModuleInput and ModulePlayer
Now if you press a key before being in a corner, ModulePlayer keep the
last key and if different to the direction, this change of direction.

Example:
If you're down and you're in the middle of the wall to 1 second of
reaching the corner and press the key to turn right, the program will
save you've pressed the turn. When you get there if you have not pressed
another key, the address to be changed.